+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## 2022/10/14 - Version 0.04
+### Features Added
+#### Policy Objects
+* Address Groups - Delete Edit
+
 ## 2022/10/12 - Version 0.03
 ### Features Added
 #### Policy Objects